Confirm alert updated

+++ src/app/deletebranch/deletebranch.component.ts
@@ -44,18 +44,22 @@ export class DeletebranchComponent implements OnInit {
 
   deleteFun(id: any, revId: any) {
     try {
-      this.departmentObj.deleteFunction(id, revId).subscribe(
-        (res) => {
-          alert('Are you want to delete this branch ?');
-          window.location.reload();
-          this.toastr.success('Successfully Deletd');
-        },
-        (err: { response: { data: { errorMessage: any } } }) => {
-          let errorMessage = err.response.data.errorMessage;
-          console.error(errorMessage);
-          this.toastr.error('Unable to delete this branch');
-        }
-      );
+      const alert = confirm('Are you want to delete this branch ?');
+      if (alert == true) {
+        this.departmentObj.deleteFunction(id, revId).subscribe(
+          (res) => {
+
+            window.location.reload();
+            this.toastr.success('Successfully Deletd');
+          },
+          (err: { response: { data: { errorMessage: any } } }) => {
+            let errorMessage = err.response.data.errorMessage;
+            console.error(errorMessage);
+            console.log(errorMessage);
+            this.toastr.error('Unable to delete this branch');
+          }
+        );
+      }
     } catch (err: any) {
       console.error(err.message);
       this.toastr.error('Unable to register');
